Rename Heater file and update imports
