+++ src/index.mts
@@ -4,7 +4,7 @@ import { Server } from './components/Server.mjs'
 import { WeatherObserver } from './components/base/sensors/WeatherObserver.mjs'
 import { ElectricityManager } from './subsystems/electricity/basic/ElectricityManager.mjs'
 import { SolarPanel } from './machines/SolarPanel.mjs'
-import { Heater } from './machines/Hater.mjs'
+import { Heater } from './machines/Heater.mjs'
 import { TemperatureManager } from './subsystems/temperature/basic/TemperatureManager.mjs'
 import { Environment } from './components/base/EnclosedSpace.mjs'
 import { TemperatureExchange } from './subsystems/temperature/basic/TemperatureExchain.mjs'
